dsc: Rework `IdList<T>` for better type safety and correctness
The existing implementation required passing the type of the handle as a
separate template argument, potentially allowing incorrect handles to be
used as keys. This new implementation deduces the handle type from the
type that is stored in the list.

Additionally, because the underlying storage was simply a vector of `T`
objects, when removing an item, the existing implementation did not call
its destructor until another item was put in its place or until the list
was destructed, in order to avoid calling the destructor twice (once on
removal, and another one upon the destruction of the underlying vector,
which could lead to unexpected results, such as double freeing). This
behaviour, however, meant that resources could potentially be held on to
longer than needed in case the object's `Clear()` function didn't release
them. This is now avoided by using opaque storage as the backing storage,
and correctly destructing a value when it is removed.

As a bonus point, the new implementation uses slightly less memory, as
it doesn't need a separate free list to track the state of free items.

This was also an opportunity to get rid of the bespoke `.MoveSelfInto()`
and `.DeepCopyInto()` methods, and replace them with proper copy and
move assignment operators.

Moreover, even though references may still be invalidated upon insertion,
the iterator is now still valid as long as items were inserted after it.
While not part of this commit, this could be used to reduce the number
of places that use the weird index-based access that the list provides
and convert them to iterator use (such as in the case of the loops in
boolean.cpp, which cannot use ranged-based for loops, but can now in
theory use the iterator instead of accessing by index).

In order to make access by index more evident, however, I did remove the
subscript operator and replaced all of the code that used it with an
explicit call to `.Get()`. Note that there are still places that cannot
be converted to use iterators (essentially all of the loops marked with
`#pragma omp parallel`), but we can at least have fewer code paths that
(ab)use this feature.