file: rework saving and loading for correctness and to avoid copying
Rework the object attributes to avoid needlessly iterating over all of
them when needing to save a specific object or load a specific attribute
of an object.

While at it, add type verification for member serialisation. This caught
an issue with the `Group::forceToMesh` member, which was incorrectly
serialised as an `int`, rather than as a `bool`.

Also change the file saving logic so that it doesn't have to go through
the `sv` member of `SolveSpaceUI`, which needlessly requires copying
objects over and over.

This leaves `sv` to be used only as a temporary during file loading, but
there's no reason to keep it as a member for this, so it is now made a
local variable in the two functions that need it.