dsc: rework `List<T>` for better encapsulation and reduced maintenance
Make `List<T>` a thin wrapper around `std::vector<T>` in order to avoid
the need to maintain code that deals with raw memory management, and get
move construction and assignment for free along the way.

As part of it, remove the `n` member and export the size using a getter
function, fixing some suspicious modifications of `n` in src/mesh.cpp.

The `Size()`, `ReserveMore()`, and `RemovedLast()` methods should really
be returning and accepting `size_t`, respectively, rather than `int`,
but they were left as is in order to avoid disturbing too much code.