dsc: remove insertion by pointer from `List<T>` and `IdList<T>`
This is done to make it clearer to the reader of the code if a deep copy
or a move was intended, and while at it fix all the places where the old
pointer functions were used.

Additionally, make sure all assignments where `List<T>` and `IdList<T>`
were involved stay shallow by moving where necessary instead of copying.

After this, the usefulness of `List<T>` as a simple thin wrapper around
`std::vector<T>` is questionable, and it can technically be removed and
fully replaced by `std::vector<T>`, with additional utility functions
for replicating the functionality offered by the `ClearTags()` and
`RemoveTagged()` member functions.