general: remove `List<T>::NextAfter()` and replace usages with iterators
The `NextAfter()` iteration is more verbose and slower than simply using
range-based for loops (or indexed for loops where a range-based for loop
cannot be used).

This is mostly a mechanical change of finding and replacing things (and
adding a comment explaining why an indexed for loop was used, respectively),
but it also converts many places where mutable pointers were used to
const-qualified ones in order to better convey the intent of the iteration.

This last point also required adding const qualifiers on a few member
functions and parameters to functions.